pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["srivatsan177 <ksrivatsan177@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.95.2"}
mongoengine = "^0.27.0"
python-dotenv = "^1.0.0"
cryptography = "^40.0.2"
black = "^23.3.0"
pyjwt = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
